Add user authentication tests
