+++ server/babel.config.js
@@ -1,0 +1,18 @@
+/*
+code provided by ChatGPT4 via chat.openai.com
+
+Prompt(s): requested help configuring Jest for .mjs files
+*/
+
+module.exports = {
+    presets: [
+        [
+            '@babel/preset-env',
+            {
+                targets: {
+                    node: 'current',
+                },
+            },
+        ],
+    ],
+};
