+++ server/package.json
@@ -1,10 +1,10 @@
 {
   "name": "cs467_job_tracker",
   "version": "1.0.0",
-  "main": "server.js",
+  "main": "server.mjs",
   "scripts": {
     "test": "jest",
-    "start": "node server.js"
+    "start": "node server.mjs"
   },
   "repository": {
     "type": "git",
@@ -19,10 +19,17 @@
   "homepage": "https://github.com/derussoj/cs467_job_tracker#readme",
   "description": "Welcome to the Job Tracker backend! This Node.js application provides the server-side functionality for the Job Tracker web application.",
   "devDependencies": {
-    "jest": "^29.7.0"
+    "@babel/core": "^7.23.9",
+    "@babel/preset-env": "^7.23.9",
+    "babel-jest": "^29.7.0",
+    "jest": "^29.7.0",
+    "nock": "^13.5.3",
+    "supertest": "^6.3.4"
   },
   "dependencies": {
+    "dotenv": "^16.4.4",
     "express": "^4.18.2",
+    "express-session": "^1.18.0",
     "mongoose": "^8.1.1",
     "passport": "^0.7.0",
     "passport-github2": "^0.1.12",
